Update Module_4_Lab.md
The link to download the Microsoft Online Services Sign-In Assistant for IT Professionals RTW (https://www.microsoft.com/en-us/Download/details.aspx?id=28177). no longer works. I added the link that does work and instructions to download and install the Microsoft Online Services Sign-In Assistant for IT Professionals RTW. 

It may be worth considering adding the installer to the lab files in case this link get deprecated as well.